third_party/blink/web_tests/platform/generic/fast/dom/HTMLImageElement/image-srcset-react-to-media-changes-px-based-sizes-expected.txt

PASS img.currentSrc.indexOf("blue_rect") is not -1
PASS img.currentSrc.indexOf("image-set-4x") is not -1
PASS successfullyParsed is true

TEST COMPLETE
